+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 * expose balanced method in public API (preferred name)
 * optional brackets
 * `interator?` function and detection if iterators from `repr` and `type` [#51](https://github.com/jcubic/lips/issues/51)
+* `add-repr!` function to add string representations to objects
 ### Bug fixes
 * fix list function [#27](https://github.com/jcubic/lips/issues/27)
 * fix `(begin (gensym))` macro that evaluate to symbol